still working

+++ src/main/java/view/MessageShow.java
@@ -51,6 +51,16 @@ public class MessageShow {
             vacDescription.setVisible(true);
             this.acc = new Button("Accept Trade Request");
             this.dec = new Button("Decline Trade Request");
+        } else if (type.equals("trade request approve")){
+            this.acc = new Button("");
+            this.dec = new Button("");
+            this.dec.setVisible(false);
+            this.acc.setVisible(false);
+        }else if (type.equals("trade request decline")){
+            this.acc = new Button("");
+            this.dec = new Button("");
+            this.dec.setVisible(false);
+            this.acc.setVisible(false);
         }
         this.vacationController = vacationController1;
         this.mainController = mainController;
@@ -76,15 +86,19 @@ public class MessageShow {
 //                }*/
                 int sign = messageI.indexOf(" ");
                 String vacIdString = messageI.substring(0, sign);
-                try {
-                    vacationController1.add_message(from, vacIdString + " Your request to buy the vacation from " + vacationController1.username + " has been approved, you can start a chat with the seller to set up the payment",
-                            "request approve");
-                    acc.setDisable(true);
-                    dec.setDisable(true);
-                } catch (Exception e) {
+                if (!vacationController.is_messg_Exist(vacationController1.username, from, vacIdString + " Your request to buy the vacation from " + vacationController1.username + " has been approved,\nyou can start a chat with the seller to set up the payment")) {
+                    try {
+                        vacationController1.add_message(from, vacIdString + " Your request to buy the vacation from " + vacationController1.username + " has been approved,\nyou can start a chat with the seller to set up the payment",
+                                "request approve");
+                        acc.setDisable(true);
+                        dec.setDisable(true);
+                    } catch (Exception e) {
+                        showAlert("Ooops you have already accepted this request");
+                    }
+                    vacationController1.deleteMessage(from, vacationController1.username, messageI);
+                }else{
                     showAlert("Ooops you have already accepted this request");
                 }
-                vacationController1.deleteMessage(from, vacationController1.username, messageI);
             } else if (type.equals("request approve")) {
                 showAlert("Sorry... The site is under construction");
 
@@ -117,16 +131,24 @@ public class MessageShow {
                 showAlert("You accepted the trade request.");
                 int sign = messageI.indexOf(" ");
                 String vacIdString = messageI.substring(0, sign);
-                String[] sellerVac = getVacationDetails(vacIdString);
-                try {
-                    vacationController1.add_message(from, vacIdString + " Your request to trade the vacation from " + vacationController1.username + "to " + sellerVac[0] + " has been approved.",
-                            "trade request approve");
-                    acc.setDisable(true);
-                    dec.setDisable(true);
-                } catch (Exception e) {
+                System.out.println("my requested vacation: " + messageI.substring(1, sign-1));
+                String[] sellerVac = getVacationDetails(messageI.substring(1, sign-1));
+                if (!vacationController.is_messg_Exist(vacationController1.username, from, vacIdString + " Congratulations!\nYour request to trade the vacation from " + vacationController1.username + " to the destination " + sellerVac[0] + "\nhas been approved.")) {
+                    try {
+                        vacationController1.add_message(from, vacIdString + " Congratulations!\nYour request to trade the vacation from " + vacationController1.username + " to the destination " + sellerVac[0] + "\nhas been approved.",
+                                "trade request approve");
+                        acc.setDisable(true);
+                        dec.setDisable(true);
+                    } catch (Exception e) {
+                        showAlert("Ooops you have already accepted this request");
+                    }
+                }else{
                     showAlert("Ooops you have already accepted this request");
                 }
                 vacationController1.deleteMessage(from, vacationController1.username, messageI);
+            }
+            else if(type.equals("trade request approve")){
+
             }
 
         });
@@ -134,6 +156,25 @@ public class MessageShow {
             showAlert("Such a loss");
             acc.setDisable(true);
             dec.setDisable(true);
+            if(type.equals("trade request vacation")){
+                int sign = messageI.indexOf(" ");
+                String vacIdString = messageI.substring(0, sign);
+                System.out.println("my requested vacation: " + messageI.substring(1, sign-1));
+                String[] sellerVac = getVacationDetails(messageI.substring(1, sign-1));
+                if (!vacationController.is_messg_Exist(vacationController1.username, from, vacIdString + " We are so sorry, \n your request to trade the vacation from " + vacationController1.username + "to the destination " + sellerVac[0] + "\nhas been declined.")) {
+                    try {
+                        vacationController1.add_message(from, vacIdString + " We are so sorry, \n your request to trade the vacation from " + vacationController1.username + "to the destination " + sellerVac[0] + "\nhas been declined.",
+                                "trade request decline");
+                        acc.setDisable(true);
+                        dec.setDisable(true);
+                    } catch (Exception e) {
+                        showAlert("Ooops you have already declined this request");
+                    }
+                }else{
+                    showAlert("Ooops you have already declined this request");
+                }
+                vacationController1.deleteMessage(from, vacationController1.username, messageI);
+            }
         });
         vacDescription.setOnAction(event -> {
             if (type.equals("trade request vacation")) {
